nba_live_app/requirements.txt
gradio
kafka-python
groq
huggingface-hub<0.23.0
